refactor(flashcard-session): rename flashcard/ to session-studying/
flashcard-session/ is flashcard-only now, so a flashcard/ subfolder was
a stutter. Renamed it to session-studying/, a sibling of
session-summary/ (both phase-named). Carved out session-studying/card/
for the pieces specific to rendering/editing the single active card
(card-stage.vue, study-card.vue, study-card-edit.vue,
primed-grade-context.ts); the phase view, session-progress.vue,
study-edit-footer.vue, and rating-buttons/ stay at the session-studying/
level as session-level chrome rather than card-specific.
